Add Exts code generator.
This translates typechecking to a set of haskell functions wrapped in a
haskell module. Based on haskell-src-exts package for Haskell syntax.

This is a partial commit - typechecking is not fully functional yet. But
the basis for generating compiled functions representing types and
code for checking the sorts of types exists. Also can generate a 'main'
function that forces evaluation of toplevel sort checking functions.