Rewrote the parser.
Parser now conforms to a formal grammar. It is as a result much
shorter and simpler. It is written in applicative style, rather than
monadic style, because applicative style is sufficient, and because it
opens the door to transitioning to a different parser combinator
library than parsec with better sharing and performance.