Remove usage of 'text' package for utf-8 support.
It was being a maintainance and a performance problem. New text-0.5
causes random segfaults so we had to pin the version to text==0.4.
Another problem is that measure-complexity.sh uncovered super
loglinear complexity. If we replace text with pure bytestrings, we
regain loglinear complexity, in addition to a hefty constant factor
improvement in performance. :)