Switch to a bidirectional core representation.
This means that lambda abstractions are domain free, because we only
ever check lambdas. We don't infer their type. Thus type checkable
terms and code now have the same representation. This means that we
could conceivably drop the Code data type in the runtime and only have
one representation of terms.