Plug anf and closure conversion passes into compilation pipeline.
We have a generic 'pass' combinator lifting a transformation to the
'DkM' monad. Each pass as well as checking and code generation have
been annotated with an SCC pragma to facilitate profiling. The runtime
has been modified to handle let forms and we remove UBox'es, because
they are no longer required thanks to anf.