Generated guards were wrong.
In Runtime.hs, reflect was not exported, but it is quite useful to have
it to check wether a Prop is true or false (I still do not see the point
of having prop while we could simply use Bool, convertible is something
very decidable!). The constraints generated by CodeGen/Exts.hs must use
this reflect call, otherwise, the resulting Haskell file won't
typecheck.