Attempted to clean the runtime system.
UBox are no longer terms, we use pairs of terms and code. The
bidirectionnal typechecker is now supposed to work correctly. Type
annotations can be added on lambda terms, they are used by the synthesis
process. This implies massive changes in the code generator.

I also removed the Prop thing by replacing it with (). It should be OK.