Massive rework on code generation.
We keep track of annotations on lambdas when there is one in the source
dk. The "box" postfixed variables are now gone, terms are postfixed by
"t", codes are postfixed by "c", that's the only rule. UBox is gone, so
we must generate a "Pair" value at application nodes, and match over
Pair values in term abstractions.

transformM is now a special case of descendM.

Some factoring would be welcome in the Exts.hs file when dealing with
names.

The next step is to remove the whole annotation mechanism which is now
obsolete (we only need annotations on lambda abstractions).